requirements/base.txt
django
psycopg2-binary
python-environ
django-cors-headers
django-jazzmin
drf-yasg
djangorestframework
django-filter
django-rest-authtoken-2.1.4
djangorestframework-simplejwt-5.2.2
pyjwt-2.8.0

